Split up utils.misc.
