+++ aslprep/tests/test_utils_cbf.py
@@ -1,17 +1,17 @@
-"""Tests for the aslprep.utils.misc module."""
+"""Tests for the aslprep.utils.cbf module."""
 import numpy as np
 
-from aslprep.utils import misc
+from aslprep.utils import cbf
 
 
 def test_estimate_att_pcasl_3d():
-    """Smoke test aslprep.utils.misc.estimate_att_pcasl with 3D data (no slice timing)."""
+    """Smoke test aslprep.utils.cbf.estimate_att_pcasl with 3D data (no slice timing)."""
     n_voxels, n_unique_plds = 1000, 6
     base_plds = np.linspace(0.25, 1.5, n_unique_plds)[None, :]
     plds = np.repeat(base_plds, n_voxels, axis=0)
     deltam_arr = np.random.random((n_voxels, n_unique_plds))
     tau = np.full(n_unique_plds, 1.4)
-    att = misc.estimate_att_pcasl(
+    att = cbf.estimate_att_pcasl(
         deltam_arr=deltam_arr,
         plds=plds,
         lds=tau,
@@ -22,7 +22,7 @@ def test_estimate_att_pcasl_3d():
 
 
 def test_estimate_att_pcasl_2d():
-    """Smoke test aslprep.utils.misc.estimate_att_pcasl with slice-shifted PLDs."""
+    """Smoke test aslprep.utils.cbf.estimate_att_pcasl with slice-shifted PLDs."""
     n_voxels, n_unique_plds, n_slice_times = 1000, 6, 50
     base_plds = np.linspace(0.25, 1.5, n_unique_plds)
     slice_times = np.linspace(0, 3, n_slice_times)
@@ -30,7 +30,7 @@ def test_estimate_att_pcasl_2d():
     plds = np.repeat(slice_shifted_plds, n_voxels // n_slice_times, axis=0)
     deltam_arr = np.random.random((n_voxels, n_unique_plds))
     tau = np.full(n_unique_plds, 1.4)
-    att = misc.estimate_att_pcasl(
+    att = cbf.estimate_att_pcasl(
         deltam_arr=deltam_arr,
         plds=plds,
         lds=tau,
@@ -41,7 +41,7 @@ def test_estimate_att_pcasl_2d():
 
 
 def test_estimate_cbf_pcasl_multipld():
-    """Smoke test aslprep.utils.misc.estimate_cbf_pcasl_multipld with slice-shifted PLDs."""
+    """Smoke test aslprep.utils.cbf.estimate_cbf_pcasl_multipld with slice-shifted PLDs."""
     n_voxels, n_unique_plds, n_slice_times = 1000, 6, 50
     base_plds = np.linspace(0.25, 1.5, n_unique_plds)
     slice_times = np.linspace(0, 3, n_slice_times)
@@ -50,7 +50,7 @@ def test_estimate_cbf_pcasl_multipld():
     deltam_arr = np.random.random((n_voxels, n_unique_plds))
     scaled_m0data = np.random.random(n_voxels)
 
-    misc.estimate_cbf_pcasl_multipld(
+    cbf.estimate_cbf_pcasl_multipld(
         deltam_arr=deltam_arr,
         scaled_m0data=scaled_m0data,
         plds=plds,
